phpunit: User model

+++ app/User.php
@@ -6,6 +6,17 @@ use Illuminate\Foundation\Auth\User as Authenticatable;
 
 use Hash;
 
+
+/**
+ * 
+ * @method assignRole(int|string|object $role)
+ * @method deleteRole(int|string|object $role)
+ * 
+ * @static @method add(array $arrData)
+ * @static @method getModel(int|string|object $user)
+ * @static @method checkArrayUser(array $arrData) 
+ * 
+ */
 class User extends Authenticatable
 {
     /**
@@ -35,65 +46,76 @@ class User extends Authenticatable
 	
 	
 	/**
-	 * Привязывание роли к пользователю
+	 * Assign role to the user
+	 * 
+	 * @param {int|string|object}
+	 * 	@param {int} id role
+	 * 	@param {string} name role
+	 * 	@param {object} object role
+	 *
+	 * @returns {object}
 	 */
-	public function assignRole($nameRole = '')
+	public function assignRole($role = '')
 	{
-		if($nameRole != ''){
-			$roleModel = new Role;
-			$role = $roleModel->getByName($nameRole);
-			
-			if(count($role) > 0){
-				$check = $this->roles()->find($role->id);
-				if(count($check) == 0){
-					$this->roles()->save($role);
-				}
-			}
-			
+		$role = Role::getModel($role);
+		
+		if(!is_null($role)){
+			$this->roles()->save($role);
 		}
+
 		return $this;
 	}
 	
 	
 	
 	/**
-	 * Отвязывание роли к пользователю
+	 * Delete role to the user
+	 * 
+	 * @param {int|string|object}
+	 * 	@param {int} id role
+	 * 	@param {string} name role
+	 * 	@param {object} object role
+	 *
+	 * @returns {object}
 	 */
-	public function deleteRole($nameRole = '')
+	public function deleteRole($role = '')
 	{
-		if($nameRole != ''){
-			$roleModel = new Role;
-			$role = $roleModel->getByName($nameRole);
-			
-			if(!is_null($role)){
-				$this->roles()->where('id', $role->id)->delete();
-			}
-			
+		$role = Role::getModel($role);
+		
+		if(!is_null($role)){
+			$this->roles()->detach($role);
 		}
+		
 		return $this;
 	}
 	
 	
 	
 	/**
-	 * Изменение пользователя
+	 * Save user
+	 *
+	 * @param {array} $arrData array with data to be added
+	 * 	@param {string} $arrData['name'] name user
+	 * 	@param {string} $arrData['email'] email user
+	 * 	@param {string} $arrData['password'] password user
+	 *
+	 * @returns {object|null}
+	 *
 	 */
-	public static function updateData($arrParam)
+	public static function add($arrData = array())
 	{
-		if(is_array($arrParam) === false){
-			return false;
+		if(self::checkArrayUser($arrData) === false){
+			return null;
 		}
 		
 		$arrDefault = array(
-			'id' => '',
 			'name' => '',
 			'email' => '',
 			'password' => '',
 		);
-		$res = array_merge($arrDefault, $arrParam);
+		$res = array_merge($arrDefault, $arrData);
 		
 		$user = new User;
-		if($res['id'] != ''){$user = User::find($res['id']);}
 		$user->name = $res['name'];
 		$user->email = $res['email'];
 		if($res['password'] != ''){
@@ -107,14 +129,72 @@ class User extends Authenticatable
 	
 	
 	/**
-	 * Удаление пользователя
+	 * Delete user
+	 *
+	 * @param {int|string|object}
+	 * 	@param {int} id user
+	 * 	@param {string} email user
+	 * 	@param {object} object user
+	 *
+	 * @returns {true}
 	 */
-	public static function deleteData($id)
+	public static function del($user = '')
 	{
-		$user = User::find($id);
-		$user->delete();
+		$user = self::getModel($user);
+		if(!is_null($user)){
+			$user->delete();
+		}
 		
 		return true;
 	}
-			
+	
+	
+	
+	
+	/**
+	 *  Getting the user by id or email or object
+	 * 
+	 * @param {int|string|object}
+	 * 	@param {int} id user
+	 * 	@param {string} email user
+	 * 	@param {object} object user
+	 *
+	 * @returns {object|null}
+	 */
+	public static function getModel($user = '')
+	{
+		$userModel = null;
+		if(is_string($user) === true){
+			$userModel = self::where('email', $user);
+		}
+		if(is_int($user) === true){
+			$userModel = self::find($user);
+		}
+		if(is_object($user) === true){
+			$userModel = self::find($user->id);
+		}
+		
+		return $userModel;
+	}
+	
+	
+	
+	
+	/**
+	 * Checking to add to the array user
+	 * 
+	 * @returns {true|false}
+	 */
+	static function checkArrayUser($arrData = array())
+	{
+		if(
+			is_array($arrData) === true 
+			&& array_key_exists('name', $arrData) === true
+			&& array_key_exists('email', $arrData) === true
+		){
+			return true;
+		}
+		
+		return false;
+	}	
 }
